QPID-7527: [Java Broker] ProtocolOutputConverterImpl should cast operands to long to avoid truncation to int

+++ broker-plugins/amqp-0-8-protocol/src/main/java/org/apache/qpid/server/protocol/v0_8/ProtocolOutputConverterImpl.java
@@ -462,7 +462,7 @@ public class ProtocolOutputConverterImpl implements ProtocolOutputConverter
 
         public long getSize()
         {
-            return OVERHEAD + _methodBody.getSize() + _headerBody.getSize() + _contentBody.getSize();
+            return OVERHEAD + (long)_methodBody.getSize() + (long)_headerBody.getSize() + (long)_contentBody.getSize();
         }
 
         @Override
@@ -510,7 +510,7 @@ public class ProtocolOutputConverterImpl implements ProtocolOutputConverter
 
         public long getSize()
         {
-            return OVERHEAD + _methodBody.getSize() + _headerBody.getSize() ;
+            return OVERHEAD + (long)_methodBody.getSize() + (long)_headerBody.getSize() ;
         }
 
         @Override
